+++ crates/fuzz/src/lib.rs
@@ -5,3 +5,4 @@ pub mod fuzzing_stats;
 pub mod program_test_client_blocking;
 pub mod snapshot;
 pub type AccountId = u8;
+pub mod trident_accounts_struct;